% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{binomial_price}
\alias{binomial_price}
\title{Cox-Ross-Rubinstein binomial option price}
\usage{
binomial_price(
  S,
  K,
  r,
  q = 0,
  sigma,
  T,
  n = 100,
  type = c("call", "put"),
  american = FALSE
)
}
\arguments{
\item{S}{Spot price}

\item{K}{Strike price}

\item{r}{Risk-free rate (annual, decimal)}

\item{q}{Dividend yield (annual, decimal, zero by default)}

\item{sigma}{Volatility (annual, decimal)}

\item{T}{Time to maturity (years)}

\item{n}{number of steps}

\item{type}{'call' or 'put'}

\item{american}{allowing for early exercise}
}
\value{
Option price.
}
\description{
Calculates option price based on a Cox-Ross-Rubinstein binomial tree.
}
\examples{
binomial_price(100, 100, 0.03, 0, 0.2, 0.5, 100, 'call', american = FALSE)

}
\references{
Das, S., & Sundaram, R. (2015) \emph{Derivatives: Principles and Practice}. McGraw-Hill Education.
}
