% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeks.R
\name{bs_greeks}
\alias{bs_greeks}
\title{Black-Scholes option Greeks}
\usage{
bs_greeks(S, K, r, q = 0, sigma, T, type = c("call", "put"))
}
\arguments{
\item{S}{Spot price}

\item{K}{Strike price}

\item{r}{Risk-free rate (annual, decimal)}

\item{q}{Dividend yield (annual, decimal, zero by default)}

\item{sigma}{Volatility (annual, decimal)}

\item{T}{Time to maturity (years)}

\item{type}{'call' or 'put'}
}
\value{
List of option Greeks.
}
\description{
Determines option Greeks for a given set of option characteristics.
}
\examples{
bs_greeks(100, 100, 0.03, 0, 0.2, 0.5, 'call')

}
\references{
Das, S., & Sundaram, R. (2015) \emph{Derivatives: Principles and Practice}. McGraw-Hill Education.
}
