% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs.R
\name{bs_price}
\alias{bs_price}
\title{Black-Scholes European option price}
\usage{
bs_price(S, K, r, q = 0, sigma, T, type = c("call", "put"))
}
\arguments{
\item{S}{Spot price}

\item{K}{Strike price}

\item{r}{Risk-free rate (annual, decimal)}

\item{q}{Dividend yield (annual, decimal, zero by default)}

\item{sigma}{Volatility (annual, decimal)}

\item{T}{Time to maturity (years)}

\item{type}{'call' or 'put'}
}
\value{
Option price.
}
\description{
Calculates option price based on Black-Scholes formula.
}
\examples{
bs_price(100, 100, 0.03, 0, 0.2, 0.5, 'call')

}
\references{
Das, S., & Sundaram, R. (2015) \emph{Derivatives: Principles and Practice}. McGraw-Hill Education.
}
