% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo.R
\name{mc_price}
\alias{mc_price}
\title{Monte Carlo simulated European option price}
\usage{
mc_price(
  S,
  K,
  r,
  q = 0,
  sigma,
  T,
  n,
  type = c("call", "put"),
  antithetic = TRUE,
  control_variate = TRUE,
  seed = NULL
)
}
\arguments{
\item{S}{Spot price}

\item{K}{Strike price}

\item{r}{Risk-free rate (annual, decimal)}

\item{q}{Dividend yield (annual, decimal, zero by default)}

\item{sigma}{Volatility (annual, decimal)}

\item{T}{Time to maturity (years)}

\item{n}{Number of steps}

\item{type}{'call' or 'put'}

\item{antithetic}{Antithetic variance reduction method, TRUE by default}

\item{control_variate}{Control variate variance reduction method, TRUE by default}

\item{seed}{Controls randomness for reproducibility}
}
\value{
Option price.
}
\description{
Monte Carlo simulated European option price
}
\examples{
mc_price(100, 100, 0.03, 0, 0.2, 0.5, 1000, type = 'call', antithetic = TRUE, control_variate = TRUE)

}
\references{
Das, S., & Sundaram, R. (2015) \emph{Derivatives: Principles and Practice}. McGraw-Hill Education.
}
